Move scoring function to util
